# app/backend/cmd/api/docs/swagger.yaml
definitions:
  common.ErrorResponse:
    properties:
      error:
        description: 'example: user not found'
        type: string
    type: object
...
  handlers.CommentResponse:
    properties:
      created_at:
        description: 'example: 2025-10-11T14:00:00Z'
        type: string
      created_by:
        description: 'example: 2'
        type: integer
      defect_id:
        description: 'example: 1'
        type: integer
      id:
        description: 'example: 1'
        type: integer
      text:
        description: 'example: Broken glass needs replacement'
        type: string
    type: object
...
  handlers.CreateCommentRequest:
    properties:
      defect_id:
        description: 'example: 1'
        type: integer
      text:
        description: 'example: Found a broken window on 3rd floor'
        type: string
    type: object
...
  handlers.DefectAttachmentResponse:
    properties:
      defect_id:
        description: 'example: 2'
        type: integer
      id:
        description: 'example: 1'
        type: integer
      url:
        description: 'example: internal/uploads/defect_attachments/1759835216551583000_broken_wall.png'
        type: string
    type: object
...
info:
  contact: {}
  description: this is documentation for buildefect API
  title: buildefect api
  version: "1.0"
paths:
  /api/attachments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a defect attachment by attachment ID
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: attachment deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete defect attachment
      tags:
      - defect-attachments
    get:
      consumes:
      - application/json
      description: Get defect attachment by attachment ID
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DefectAttachmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get defect attachment
      tags:
      - defect-attachments
...
  /api/comments:
    get:
      consumes:
      - application/json
      description: Get all comments for a specific defect
      parameters:
      - description: Defect ID
        in: query
        name: defect_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.CommentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create a new comment for a specific defect. Requires authentication.
      parameters:
      - description: Comment payload
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - comments
  /api/comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment by ID. Only users with role "observer" are allowed.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted comment with id {id}
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
    get:
      consumes:
      - application/json
      description: Get a comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get a comment
      tags:
      - comments
...
  /api/defects/{id}/attachments:
    get:
      consumes:
      - application/json
      description: Get all attachments for a specific defect
      parameters:
      - description: Defect ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.DefectAttachmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List defect attachments
      tags:
      - defect-attachments
    post:
      consumes:
      - multipart/form-data
      description: Upload a file for a specific defect. Requires authentication.
      parameters:
      - description: Defect ID
        in: path
        name: id
        required: true
        type: integer
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.DefectAttachmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload defect attachment
      tags:
      - defect-attachments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"